Shared / OkResponse - edited statusCode parameter

+++ Microservices/Catalog/NET5Academy.Services.Catalog/Services/CategoryService.cs
@@ -28,38 +28,38 @@ namespace NET5Academy.Services.Catalog.Services
         {
             var categories = await _categoryCollection.Find(x => true).ToListAsync();
             var mapDtos = _mapper.Map<List<CategoryDto>>(categories);
-            return OkResponse<List<CategoryDto>>.Success((int)HttpStatusCode.OK, mapDtos);
+            return OkResponse<List<CategoryDto>>.Success(HttpStatusCode.OK, mapDtos);
         }
 
         public async Task<OkResponse<CategoryDto>> GetByIdAsync(string id)
         {
             if (string.IsNullOrEmpty(id))
             {
-                return OkResponse<CategoryDto>.Error((int)HttpStatusCode.BadRequest, "Id cannot be empty!");
+                return OkResponse<CategoryDto>.Error(HttpStatusCode.BadRequest, "Id cannot be empty!");
             }
 
             var category = await _categoryCollection.Find(x => x.Id == id).FirstOrDefaultAsync();
             if(category == null)
             {
-                return OkResponse<CategoryDto>.Error((int)HttpStatusCode.NotFound, "Category not found!");
+                return OkResponse<CategoryDto>.Error(HttpStatusCode.NotFound, "Category not found!");
             }
 
             var mapDto = _mapper.Map<CategoryDto>(category);
-            return OkResponse<CategoryDto>.Success((int)HttpStatusCode.OK, mapDto);
+            return OkResponse<CategoryDto>.Success(HttpStatusCode.OK, mapDto);
         }
 
         public async Task<OkResponse<CategoryDto>> CreateAsync(string name)
         {
             if (string.IsNullOrEmpty(name))
             {
-                return OkResponse<CategoryDto>.Error((int)HttpStatusCode.BadRequest, "Name cannot be empty!");
+                return OkResponse<CategoryDto>.Error(HttpStatusCode.BadRequest, "Name cannot be empty!");
             }
 
             var category = new Category { Name = name };
             await _categoryCollection.InsertOneAsync(category);
             
             var mapDto = _mapper.Map<CategoryDto>(category);
-            return OkResponse<CategoryDto>.Success((int)HttpStatusCode.OK, mapDto);
+            return OkResponse<CategoryDto>.Success(HttpStatusCode.OK, mapDto);
         }
     }
 }
